fixing update product & confirm delete

+++ src/pages/MasterData/Bundles/index.jsx
@@ -44,28 +44,45 @@ const Bundles = () => {
     }
     setIsLoading(false);
   }
-
+  
   async function onDelete(uid) {
     try {
       const result = await Swal.fire({
-        title: "Apakah anda yakin menghapus bundle ini?",
+        icon: "question",
+        title: "Apakah Anda yakin ingin menghapus bundle ini?",
         text: "Anda tidak akan dapat mengembalikannya!",
-        icon: "warning",
         showCancelButton: true,
         confirmButtonText: "Ya, Hapus",
         cancelButtonText: "Batal",
       });
 
       if (result.isConfirmed) {
-        await axios.delete(`${ApiEndpoint.BUNDLES}/${uid}`);
-        Swal.fire(
-          "Berhasil!",
-          "Anda berhasil menghapus data bundle ini.",
-          "success"
-        );
-        getDataBundle(query);
-      } else {
-        Swal.fire("Batal", "Hapus data bundle dibatalkan.", "info");
+        const { value: input } = await Swal.fire({
+          icon: "warning",
+          title: "Verifikasi",
+          text: `Silahkan ketik "hapusdata" untuk melanjutkan verifikasi hapus data !`,
+          input: "text",
+          showCancelButton: true,
+          confirmButtonText: "Konfirmasi",
+          cancelButtonText: "Batal",
+          inputValidator: (value) => {
+            if (!value || value.trim().toLowerCase() !== "hapusdata") {
+              return 'Anda harus memasukkan kata "hapusdata" untuk melanjutkan verifikasi hapus data!';
+            }
+          },
+        });
+
+        if (input && input.trim().toLowerCase() === "hapusdata") {
+          await axios.delete(`${ApiEndpoint.BUNDLES}/${uid}`);
+          Swal.fire(
+            "Berhasil!",
+            "Anda berhasil menghapus data bundle ini.",
+            "success"
+          );
+          getDataBundle(query);
+        } else {
+          Swal.fire("Batal", "Hapus data bundle dibatalkan.", "info");
+        }
       }
     } catch (err) {
       Swal.fire("Gagal", err.response.data.message, "error");
@@ -256,7 +273,7 @@ const Bundles = () => {
                                     <Icon icon="heroicons:eye" />
                                   </button>
                                 </Tooltip>
-                                <Tooltip
+                                {/* <Tooltip
                                   content="Edit"
                                   placement="top"
                                   arrow
@@ -286,7 +303,7 @@ const Bundles = () => {
                                   >
                                     <Icon icon="heroicons:trash" />
                                   </button>
-                                </Tooltip>
+                                </Tooltip> */}
                               </div>
                             </td>
                           </tr>
