+++ src/pages/MasterDivisi/MasterDealer/index.jsx
@@ -77,24 +77,41 @@ const Dealers = () => {
   async function onDelete(uid) {
     try {
       const result = await Swal.fire({
-        title: "Apakah anda yakin menghapus dealer ini?",
+        icon: "question",
+        title: "Apakah Anda yakin ingin menghapus dealer ini?",
         text: "Anda tidak akan dapat mengembalikannya!",
-        icon: "warning",
         showCancelButton: true,
         confirmButtonText: "Ya, Hapus",
         cancelButtonText: "Batal",
       });
 
       if (result.isConfirmed) {
-        await axios.delete(`${ApiEndpoint.DEALER}/${uid}`);
-        Swal.fire(
-          "Berhasil!",
-          "Anda berhasil menghapus data dealer ini.",
-          "success"
-        );
-        getDataDealer(query);
-      } else {
-        Swal.fire("Batal", "Hapus data dealer dibatalkan.", "info");
+        const { value: input } = await Swal.fire({
+          icon: "warning",
+          title: "Verifikasi",
+          text: `Silahkan ketik "hapusdata" untuk melanjutkan verifikasi hapus data !`,
+          input: "text",
+          showCancelButton: true,
+          confirmButtonText: "Konfirmasi",
+          cancelButtonText: "Batal",
+          inputValidator: (value) => {
+            if (!value || value.trim().toLowerCase() !== "hapusdata") {
+              return 'Anda harus memasukkan kata "hapusdata" untuk melanjutkan verifikasi hapus data!';
+            }
+          },
+        });
+
+        if (input && input.trim().toLowerCase() === "hapusdata") {
+          await axios.delete(`${ApiEndpoint.DEALER}/${uid}`);
+          Swal.fire(
+            "Berhasil!",
+            "Anda berhasil menghapus data dealer ini.",
+            "success"
+          );
+          getDataDealer(query);
+        } else {
+          Swal.fire("Batal", "Hapus data dealer dibatalkan.", "info");
+        }
       }
     } catch (err) {
       Swal.fire("Gagal", err.response.data.message, "error");
@@ -442,7 +459,7 @@ const Dealers = () => {
               </div>
               <div className="text-base text-slate-600 dark:text-slate-300 mb-4">
                 <Textarea
-                  label="Alamat Cabang"
+                  label="Alamat Dealer"
                   id="pn4"
                   rows="4"
                   placeholder="Masukkan alamat dealer"
